% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRsimulator-package.R
\docType{package}
\name{PCRsimulator}
\alias{PCRsimulator}
\alias{PCRsimulator-package}
\title{PCR simulator of for forensic genotyping}
\description{
PCRsimulator
}
\details{
A PCR simulator of for forensic genotyping
}
\author{
James Curran
}
